# docs/design/openapi.yaml
openapi: 3.1.0
info:
  title: BangMusic Server API
  version: 1.0.0
  summary: 자가호스팅 BangMusic 음악 서버와 모바일 앱 사이의 계약
  description: |
    설계 초안(2026-10-02). 구현 전 계약이며 동작하는 서버가 아직 없다.

    공통 규칙
    - 모든 경로는 `/v1` 아래에 있다. 호환되지 않는 변경은 `/v2`로만 낸다.
    - `/v1` 안에서는 추가만 한다(새 필드·새 엔드포인트·새 enum 값). 앱은 모르는 필드와 enum 값을 무시해야 한다.
    - 서버가 무엇을 지원하는지는 `GET /v1/server`의 `api.minor`와 `features`로 판단한다.
    - 오류는 RFC 9457 `application/problem+json`이며 기계 판독용 `code`를 포함한다.
    - 응답에는 서버의 파일 경로가 절대 포함되지 않는다. 모든 자원은 불투명 ID로만 가리킨다.
    - 시간은 RFC 3339 UTC 문자열, 길이는 밀리초 정수다.
    - 권한이 없는 자원은 존재 여부를 숨기기 위해 404 `not_found`로 응답한다. 403은 역할이 부족한 관리 API에만 쓴다.
    - 모든 응답에 `X-Request-Id`가 있다. 앱은 오류 보고에 이 값을 표시한다.
    - 모든 동작은 `426 client_too_old`(앱 API minor가 서버 최소값보다 낮음)와 `503`(`setup_required`, `maintenance` 등)을 돌려줄 수 있다.
  license:
    name: 미정 (서버 패키지 라이선스는 제품 소유자 결정 사항)
    identifier: LicenseRef-TBD
servers:
  - url: https://{host}/v1
    description: 구매자가 운영하는 서버. 앱은 사용자가 입력한 주소를 사용한다.
    variables:
      host:
        default: music.example.net
...
security:
  - bearerAuth: []
paths:
...
  /me:
    get:
      tags: [me]
      operationId: getMe
      summary: 본인 정보와 접근 가능한 라이브러리
      responses:
        '200':
          description: 본인 정보
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Me' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '426': { $ref: '#/components/responses/ClientTooOld' }
        '503': { $ref: '#/components/responses/Unavailable' }
    delete:
      tags: [me]
      operationId: deleteMe
      summary: 본인 계정과 개인 데이터 삭제
      description: |
        플레이리스트, 재생 기록, 가사 보정, 세션을 삭제한다. 음악 원본은 건드리지 않는다.
        마지막 관리자 계정은 삭제할 수 없다(409 `last_admin`).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties:
                password: { type: string }
      responses:
        '204': { description: 삭제됨 }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/Conflict' }
        '426': { $ref: '#/components/responses/ClientTooOld' }
        '503': { $ref: '#/components/responses/Unavailable' }
...
  /me/export:
    get:
      tags: [me]
      operationId: exportMe
      summary: 본인 데이터 내보내기 (플레이리스트·재생 기록 JSON)
      responses:
        '200':
          description: 내보내기 문서
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '426': { $ref: '#/components/responses/ClientTooOld' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /home:
    get:
      tags: [catalog]
      operationId: getHome
      summary: 홈 화면용 묶음 (플레이리스트, 최근 들은 곡, 많이 들은 곡, 최근 추가)
      parameters:
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 50, default: 20 } }
      responses:
        '200':
          description: 홈 구획
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Home' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '426': { $ref: '#/components/responses/ClientTooOld' }
        '503': { $ref: '#/components/responses/Unavailable' }
...
  /tracks/{trackId}/lyrics:
    get:
      tags: [lyrics]
      operationId: getLyrics
      summary: 원문·한글 발음·한국어 번역 가사
      parameters:
        - { $ref: '#/components/parameters/TrackId' }
      responses:
        '200':
          description: 가사 묶음. 가사가 하나도 없으면 `variants`가 빈 배열이다(404가 아님).
          headers:
            ETag: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Lyrics' }
              example:
                track_id: trk_01JB2P3C5E7G9J1L3N5Q7S9U1W
                version: 3
                offset_ms: -250
                variants:
                  - kind: original
                    language: ja
                    synced: true
                    source: { type: sidecar, name: 서버의 .lrc 파일 }
                    lines: [{ t_ms: 12300, text: 夜が明ける }, { t_ms: 17850, text: 空が染まる }]
                  - kind: pronunciation_ko
                    language: ko
                    synced: true
                    source: { type: user, name: 사용자 입력 }
                    lines: [{ t_ms: 12300, text: 요루가 아케루 }, { t_ms: 17850, text: 소라가 소마루 }]
                  - kind: translation_ko
                    language: ko
                    synced: true
                    source: { type: user, name: 사용자 입력 }
                    lines: [{ t_ms: 12300, text: 밤이 밝아 온다 }, { t_ms: 17850, text: 하늘이 물든다 }]
        '304': { description: 변경 없음 }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '426': { $ref: '#/components/responses/ClientTooOld' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /tracks/{trackId}/lyrics/{kind}:
    put:
      tags: [lyrics]
      operationId: putLyricsVariant
      summary: 가사 한 종류를 직접 입력/교체 (관리자)
      description: 일반 사용자는 403. 본문은 LRC 또는 일반 텍스트이며 서버가 파싱해 저장한다.
      parameters:
        - { $ref: '#/components/parameters/TrackId' }
        - { name: kind, in: path, required: true, schema: { $ref: '#/components/schemas/LyricsKind' } }
        - { $ref: '#/components/parameters/IfMatch' }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [format, body]
              properties:
                format: { type: string, enum: [lrc, plain] }
                language: { type: string }
                body: { type: string, maxLength: 200000 }
                source_name: { type: string }
                license_note: { type: string }
      responses:
        '200':
          description: 저장됨
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Lyrics' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '412': { $ref: '#/components/responses/PreconditionFailed' }
        '426': { $ref: '#/components/responses/ClientTooOld' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /tracks/{trackId}/lyrics/offset:
    put:
      tags: [lyrics]
      operationId: putLyricsOffset
      summary: 사용자별 가사 동기화 보정값 저장
      parameters:
        - { $ref: '#/components/parameters/TrackId' }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [offset_ms]
              properties:
                offset_ms: { type: integer, minimum: -10000, maximum: 10000 }
      responses:
        '204': { description: 저장됨 }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '426': { $ref: '#/components/responses/ClientTooOld' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /playlists:
    get:
      tags: [playlists]
      operationId: listPlaylists
      summary: 내 플레이리스트
      parameters:
        - { $ref: '#/components/parameters/Cursor' }
        - { $ref: '#/components/parameters/Limit' }
      responses:
        '200':
          description: 플레이리스트 페이지
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PlaylistPage' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '426': { $ref: '#/components/responses/ClientTooOld' }
        '503': { $ref: '#/components/responses/Unavailable' }
    post:
      tags: [playlists]
      operationId: createPlaylist
      summary: 플레이리스트 생성 (멱등 키)
      parameters:
        - { $ref: '#/components/parameters/IdempotencyKey' }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, minLength: 1, maxLength: 200 }
                description: { type: string, maxLength: 2000 }
                track_ids: { type: array, maxItems: 5000, items: { type: string } }
      responses:
        '201':
          description: 생성됨
          headers:
            ETag: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Playlist' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '426': { $ref: '#/components/responses/ClientTooOld' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /playlists/{playlistId}:
    parameters:
      - { $ref: '#/components/parameters/PlaylistId' }
    get:
      tags: [playlists]
      operationId: getPlaylist
      summary: 플레이리스트 메타데이터 조회
      responses:
        '200':
          description: 플레이리스트 (항목 제외)
          headers:
            ETag: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Playlist' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '426': { $ref: '#/components/responses/ClientTooOld' }
        '503': { $ref: '#/components/responses/Unavailable' }
    patch:
      tags: [playlists]
      operationId: updatePlaylist
      summary: 이름·설명 수정
      parameters:
        - { $ref: '#/components/parameters/IfMatch' }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string, minLength: 1, maxLength: 200 }
                description: { type: string, maxLength: 2000 }
      responses:
        '200':
          description: 수정됨
          headers:
            ETag: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Playlist' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '412': { $ref: '#/components/responses/PreconditionFailed' }
        '428': { $ref: '#/components/responses/PreconditionRequired' }
        '426': { $ref: '#/components/responses/ClientTooOld' }
        '503': { $ref: '#/components/responses/Unavailable' }
    delete:
      tags: [playlists]
      operationId: deletePlaylist
      summary: 플레이리스트 삭제 (멱등)
      responses:
        '204': { description: 삭제됨 (멱등) }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '426': { $ref: '#/components/responses/ClientTooOld' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /playlists/{playlistId}/items:
    parameters:
      - { $ref: '#/components/parameters/PlaylistId' }
    get:
      tags: [playlists]
      operationId: listPlaylistItems
      summary: 플레이리스트 항목 목록
      parameters:
        - { $ref: '#/components/parameters/Cursor' }
        - { $ref: '#/components/parameters/Limit' }
      responses:
        '200':
          description: 순서대로 정렬된 항목
          headers:
            ETag: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PlaylistItemPage' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '426': { $ref: '#/components/responses/ClientTooOld' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /playlists/{playlistId}/edits:
    post:
      tags: [playlists]
      operationId: editPlaylist
      summary: 항목 추가·삭제·이동을 한 번에 적용
      description: |
        `If-Match`의 버전이 현재와 같을 때만 전체 연산이 원자적으로 적용된다.
        다르면 412 `version_conflict`와 현재 ETag를 돌려주며, 앱은 최신 상태를 받아 연산을 다시 계산한다.
      parameters:
        - { $ref: '#/components/parameters/PlaylistId' }
        - { $ref: '#/components/parameters/IfMatch' }
        - { $ref: '#/components/parameters/IdempotencyKey' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PlaylistEdit' }
            example:
              ops:
                - { op: add, track_ids: [trk_01JB2P3C5E7G9J1L3N5Q7S9U1W], after_item_id: null }
                - { op: move, item_id: pli_01JB2Q5E7G9J1L3N5Q7S9U1W3Y, after_item_id: pli_01JB2Q4D6F8H0K2M4P6R8T0V2X }
                - { op: remove, item_ids: [pli_01JB2Q6F8H0K2M4P6R8T0V2X4Z] }
      responses:
        '200':
          description: 적용됨
          headers:
            ETag: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Playlist' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '412': { $ref: '#/components/responses/PreconditionFailed' }
        '428': { $ref: '#/components/responses/PreconditionRequired' }
        '426': { $ref: '#/components/responses/ClientTooOld' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /playlists/{playlistId}/cover:
    parameters:
      - { $ref: '#/components/parameters/PlaylistId' }
    put:
      tags: [playlists]
      operationId: putPlaylistCover
      summary: 플레이리스트 표지 업로드
      description: 서버가 디코딩 후 다시 인코딩하고 메타데이터(EXIF)를 제거한다.
      parameters:
        - { $ref: '#/components/parameters/IfMatch' }
      requestBody:
        required: true
        content:
          image/jpeg: { schema: { type: string, format: binary } }
          image/png: { schema: { type: string, format: binary } }
          image/webp: { schema: { type: string, format: binary } }
      responses:
        '200':
          description: 저장됨
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Playlist' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '412': { $ref: '#/components/responses/PreconditionFailed' }
        '413':
          description: '`payload_too_large` — `limits.max_cover_bytes` 초과'
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '415':
          description: '`unsupported_media_type`'
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '426': { $ref: '#/components/responses/ClientTooOld' }
        '503': { $ref: '#/components/responses/Unavailable' }
    delete:
      tags: [playlists]
      operationId: deletePlaylistCover
      summary: 표지 제거 (곡 표지 모자이크로 돌아감)
      responses:
        '204': { description: 제거됨 }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '426': { $ref: '#/components/responses/ClientTooOld' }
        '503': { $ref: '#/components/responses/Unavailable' }
...
  /history/events:
    post:
      tags: [history]
      operationId: postPlayEvents
      summary: 재생 기록 일괄 전송 (오프라인 기록 재동기화 포함)
      description: |
        각 이벤트의 `event_id`는 앱이 만든 UUID다. 이미 받은 `event_id`는 무시하고 `duplicates`에 센다.
        따라서 응답을 못 받았을 때 같은 묶음을 그대로 다시 보내도 안전하다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [events]
              properties:
                events:
                  type: array
                  minItems: 1
                  maxItems: 500
                  items: { $ref: '#/components/schemas/PlayEvent' }
      responses:
        '200':
          description: 처리 결과
          content:
            application/json:
              schema:
                type: object
                required: [accepted, duplicates, rejected]
                properties:
                  accepted: { type: integer }
                  duplicates: { type: integer }
                  rejected:
                    type: array
                    description: 곡이 없거나 권한이 없어 버린 이벤트. 앱은 재전송하지 않는다.
                    items:
                      type: object
                      required: [event_id, code]
                      properties:
                        event_id: { type: string, format: uuid }
                        code: { type: string }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '426': { $ref: '#/components/responses/ClientTooOld' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /history/recent:
    get:
      tags: [history]
      operationId: getRecentTracks
      summary: 최근 재생한 곡
      parameters:
        - { $ref: '#/components/parameters/Cursor' }
        - { $ref: '#/components/parameters/Limit' }
      responses:
        '200':
          description: 최근 들은 곡 (곡당 1회, 최근순)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TrackPage' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '426': { $ref: '#/components/responses/ClientTooOld' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /history/top:
    get:
      tags: [history]
      operationId: getTopTracks
      summary: 기간별 많이 재생한 곡
      parameters:
        - { name: window, in: query, schema: { type: string, enum: [30d, 90d, 365d, all], default: 90d } }
        - { $ref: '#/components/parameters/Limit' }
      responses:
        '200':
          description: 많이 들은 곡
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TrackPage' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '426': { $ref: '#/components/responses/ClientTooOld' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /history:
    delete:
      tags: [history]
      operationId: deleteHistory
      summary: 내 재생 기록 전체 삭제
      responses:
        '204': { description: 삭제됨 }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '426': { $ref: '#/components/responses/ClientTooOld' }
        '503': { $ref: '#/components/responses/Unavailable' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: '`Authorization: Bearer <access_token>`. 액세스 토큰은 불투명 문자열이며 서버가 해시로 조회한다.'
    mediaTicket:
      type: apiKey
      in: query
      name: mt
      description: |
        렌디션·표지 요청 전용 단기 서명 티켓. 사용자·세션·대상 자원·만료 시각에 묶인다.
        다른 API에는 쓸 수 없다. 서버와 프록시는 접근 로그에서 `mt` 값을 가린다.
  parameters:
    Cursor:
      name: cursor
      in: query
      description: 이전 응답의 `next_cursor`. 불투명 문자열이며 앱이 만들거나 해석하지 않는다.
      schema: { type: string }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    TrackId:
      name: trackId
      in: path
      required: true
      schema: { type: string }
    PlaylistId:
      name: playlistId
      in: path
      required: true
      schema: { type: string }
...
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: 마지막으로 받은 ETag. 없으면 428.
      schema: { type: string }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      description: 앱이 만든 UUID. 같은 키의 재요청은 처음 결과를 그대로 돌려준다(24시간 보존).
      schema: { type: string, format: uuid }
  responses:
    BadRequest:
      description: '`validation_failed` — `errors`에 필드별 사유'
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Unauthorized:
      description: |
        `access_expired`(갱신 후 재시도), `access_invalid`, `session_revoked`(재로그인 필요),
        `account_disabled`.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example: { type: 'about:blank', title: Unauthorized, status: 401, code: access_expired, detail: 액세스 토큰이 만료되었습니다., request_id: req_9Kd1 }
    Forbidden:
      description: '`forbidden` — 관리자 역할 필요'
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NotFound:
      description: '`not_found` — 없거나 접근 권한이 없음'
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Conflict:
      description: '`username_taken`, `last_admin` 등 상태 충돌'
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
    PreconditionFailed:
      description: '`version_conflict` — 다른 기기가 먼저 수정함. 응답 ETag가 현재 버전.'
      headers:
        ETag: { schema: { type: string } }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    PreconditionRequired:
      description: '`precondition_required` — `If-Match` 누락'
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    ClientTooOld:
      description: '`client_too_old` — 앱의 API 버전이 서버의 `min_client_api_minor`보다 낮음. 앱 업데이트 안내.'
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
    Unavailable:
      description: |
        `storage_unavailable`(음악 저장소 마운트 장애), `maintenance`(마이그레이션·복원 중),
        `setup_required`(관리자 계정 미생성). 지수 백오프로 재시도 가능.
      headers:
        Retry-After: { schema: { type: integer } }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
    Problem:
      type: object
      required: [type, title, status, code]
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        code:
          type: string
          description: 기계 판독용 오류 코드. 전체 목록은 02-api-guide.md §오류 코드.
        detail: { type: string, description: 사람이 읽는 설명. 경로·SQL·스택을 포함하지 않는다. }
        request_id: { type: string }
        retryable: { type: boolean }
        errors:
          type: array
          items:
            type: object
            required: [field, message]
            properties:
              field: { type: string }
              message: { type: string }
...
    UserSummary:
      type: object
      required: [id, username, role]
      properties:
        id: { type: string }
        username: { type: string }
        display_name: { type: string }
        role: { type: string, enum: [admin, member] }
    Me:
      allOf:
        - $ref: '#/components/schemas/UserSummary'
        - type: object
          required: [libraries]
          properties:
            libraries:
              type: array
              items:
                type: object
                required: [id, name, status]
                properties:
                  id: { type: string }
                  name: { type: string }
                  status: { type: string, enum: [online, unavailable] }
...
    TrackPage:
      type: object
      required: [items, next_cursor]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/Track' } }
        next_cursor: { type: [string, 'null'] }
...
    Home:
      type: object
      required: [playlists, recent_tracks, top_tracks, recently_added_albums]
      properties:
        playlists: { type: array, items: { $ref: '#/components/schemas/Playlist' } }
        recent_tracks: { type: array, items: { $ref: '#/components/schemas/Track' } }
        top_tracks: { type: array, items: { $ref: '#/components/schemas/Track' } }
        recently_added_albums: { type: array, items: { $ref: '#/components/schemas/Album' } }
...
    LyricsKind:
      type: string
      enum: [original, pronunciation_ko, translation_ko]
...
    Lyrics:
      type: object
      required: [track_id, version, offset_ms, variants]
      properties:
        track_id: { type: string }
        version: { type: integer }
        offset_ms: { type: integer, description: 요청한 사용자의 보정값. 양수면 가사를 늦춘다. }
        variants: { type: array, items: { $ref: '#/components/schemas/LyricsVariant' } }
    Playlist:
      type: object
      required: [id, name, version, item_count, duration_ms, updated_at]
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: [string, 'null'] }
        artwork_id: { type: [string, 'null'], description: 업로드한 표지. null이면 앱이 `mosaic_artwork_ids`로 모자이크를 만든다. }
        mosaic_artwork_ids: { type: array, maxItems: 4, items: { type: string } }
        version: { type: integer }
        item_count: { type: integer }
        duration_ms: { type: integer }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    PlaylistPage:
      type: object
      required: [items, next_cursor]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/Playlist' } }
        next_cursor: { type: [string, 'null'] }
    PlaylistItem:
      type: object
      required: [item_id, available]
      properties:
        item_id: { type: string }
        available: { type: boolean, description: false면 곡이 누락되었거나 접근 권한이 없어 `track`이 없다. }
        track: { $ref: '#/components/schemas/Track' }
        added_at: { type: string, format: date-time }
    PlaylistItemPage:
      type: object
      required: [items, next_cursor, version]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/PlaylistItem' } }
        next_cursor: { type: [string, 'null'] }
        version: { type: integer }
    PlaylistEdit:
      type: object
      required: [ops]
      properties:
        ops:
          type: array
          minItems: 1
          maxItems: 1000
          items:
            oneOf:
              - type: object
                required: [op, track_ids]
                properties:
                  op: { const: add }
                  track_ids: { type: array, minItems: 1, items: { type: string } }
                  after_item_id: { type: [string, 'null'], description: null이면 맨 앞. 생략하면 맨 뒤. }
              - type: object
                required: [op, item_ids]
                properties:
                  op: { const: remove }
                  item_ids: { type: array, minItems: 1, items: { type: string } }
              - type: object
                required: [op, item_id, after_item_id]
                properties:
                  op: { const: move }
                  item_id: { type: string }
                  after_item_id: { type: [string, 'null'] }
...
    PlayEvent:
      type: object
      required: [event_id, track_id, started_at, played_ms, completed, source]
      properties:
        event_id: { type: string, format: uuid }
        track_id: { type: string }
        started_at: { type: string, format: date-time }
        played_ms: { type: integer, minimum: 0 }
        completed: { type: boolean }
        source: { type: string, enum: [stream, offline] }
        context:
          type: object
          properties:
            type: { type: string, enum: [album, playlist, artist, search, library, queue] }
            id: { type: [string, 'null'] }
...